C backend: Don't emit code at top of procedure as #line 0
Setting ctx.linenum = 0 overrides the _emitDBG() done right above it,
so it totally doesn't make sense.

The _emitDBG() there will set ctx.linenum to the starting line of the
procedure which is ok for code at the top of the procedure, even above
the procedure's init label, and it's definitely than emitting #line 0
which isn't meaningful at all.