symbGetWstrLen(): Remove unnecessary cunsg()
Fixes a mixing signed/unsigned warning. There's not much use in doing an
unsigned division here. The coding style of fbc is to use signed integers
to store lengths, this follows that convention. Sure enough both operands
could be casted to unsigned just as well since we know the lengths/sizes
won't ever be negative.