Fix error when having to execute compiler tools repeatedly
fbcFindBin() determines whether to run compiler tools (as.exe, ld.exe, ...)
from the compiler's bin/ directory or rely on the PATH environment for
finding them. It caches the previous result to avoid re-searching on
repeated calls (which happens when compiling modules in one fbc
invocation).

It forgot to cache the "are we relying on the system PATH" boolean, thus
repeated calls could produce wrong results.

This only affects the non-standalone build, as the standalone build always
expects to find the compiler tools inside its own tree, and doesn't use
this boolean.