crt/bits/pthreadtypes.bi: Fix long int -> LONGINT confusion
The original pthreadtypes.h header uses C's long int type in many places,
(assuming a 32bit system) it must be translated as 32bit type (Integer or
Long), but definitely not FB's 64bit LongInt.

Note: The original pthreadtypes.h uses C's long long int type in some
places too, that is 64bit and needs to be translated as LongInt, so care
must be taken because this is extremely confusing either way.