datetime.bi: Make smaller by eliminating duplicates
Previously, the header's procedure declarations were split in -lang qb/fb
versions, with the main difference being that for -lang qb it used Long
instead of Integer. Now that the header was made ready for 64bit, the
-lang fb version uses Long instead of Integer too, so they can be merged.

Only the DateDiff() declaration remains different for -lang qb/fb, because
it needs to use either __LongInt or LongInt.