ast: Fix dangling pointer in astTypeIniUpdate()
astTypeIniUpdate() calls hWalk() to search for TYPEINIs in the
given tree. hWalk() does astTypeIniFlush(), deletes the TYPEINIs,
and returns the temp var initialization code (that sets up a
temp var with the data from the TYPEINI). hWalk() also updates
the original tree to do a VAR access on the new temp var, in
place of the TYPEINI.

This prevented the dangling pointer in most cases, but it failed
to handle the case when the TYPEINI is the root of the tree, and
thus has no parent that could be updated. In that case hWalk()
just deleted the TYPEINI, leaving the "tree" variable in
astTypeIniUpdate() being a dangling pointer.

To fix this, I adjusted astTypeIniUpdate() to always wrap the
tree in a fake, temporary parent node, so hWalk() always has
something to update. This way it doesn't need to worry about
having to handle any special cases.