Fix #705: Make REDIM respect STATICLOCALS
If a procedure is marked with the STATIC keyword, all variables
declared in it should be made STATIC automatically, such that
they're allocated on the heap and are preserved for later calls.

For REDIM, this was ignored, presumably because STATIC storage
was confused with static (fixed size) arrays as opposed to
dynamic (variable size) arrays. But it's ok to have STATIC arrays
no matter whether they're fixed size or dynamic, just like with
SHAREDs or locals.

QB also creates STATIC dynamic arrays if REDIM is used to create
a dynamic array inside a procedure with STATIC, so that's another
reason why FB should behave like this too.