cVarDecl(): Generate fb_ArrayRedimEx() CALL behind DECL node
because for the C backend, it's important that variables are declared
before they're used.

This problem apparently could only show up with STATIC dynamic arrays
where the dimensions were known at the point of declaration. STATIC dynamic
arrays weren't allowed anyways though before the previous commits.

The problem with the cVarDecl() code is that it uses a var_decl
variable and astNewLINK()'s to link stuff (DECL, initializer)
together, and it's only sometimes doing astAdd() to "flush" it.
However rtlArrayRedim() always does astAdd(), and it's called
between the DECL and initializer parts, so in some cases it ended
up before both due to the delayed astAdd() for DECL/initializer.

Now it's moved behind the initializer part, which should be ok
because DECL/initializer are always flushed at that point, and the
rtlArrayRedim() isn't needed before the initializer because dynamic
arrays aren't allowed to be initialized anyways.