REDIM STATIC arrays only once at declaration point, not during every call
Array declarations that declare a dynamic array but with known dimensions
normally REDIM the array implicitly to apply the initial bounds to the
dynamic array.

The same is also done for the now allowed STATIC dynamic arrays declared
with known dimensions, but it should only be done once when the STATIC
declaration is reached for the 1st time, not every time. This is the same
as for STATIC variables which also are initialized only the 1st time their
declaration is reached, but not every time.

To implement this, I adjusted cVarDecl() to wrap the fb_ArrayRedim() call
in the same kind of STATIC flag check as done for STATIC variable
initialization. In fact I adjusted that code to allow to be re-used for
the REDIM too.

That required rtlArrayRedim() and thus also rtlErrorCheck() to be adjusted
to return their AST instead of astAdd() them, so that it can be LINKed
as needed in cVarDecl().