astNewADDROF(): Fix potential use-after-free
Before this part of the code,
l = the ADDROF's lhs, potentially a noconv CONV,
t = either l->l or l, depending on whether l is a noconv CONV or not.

Thus t <> l is the sign that l was a noconv CONV and in that case t points
to that CONV's lhs allowing it to be "skipped".

By doing t = astOptimizeTree( t ) this code had a good chance to change
t's value and thus break the following t <> l check, because if t changed
then the t <> l check would always be true now.

But if previously l = t, and now t would suddenly be different from l,
then it would enter the TRUE code path and do l->l = t, and that could be
a use-after-free if the astOptimizeTree() deleted the original l/t node.

To be safe, both l/t must be updated at the same time.

I don't know whether there's any case in practice that could trigger this,
because astOptimizeTree() would have to delete the root IDX node which I
don't think it'd currently ever do (at least it looks like none of the IDX
optimizations delete the IDX, but maybe I missed something).