rtlib: Fix & simplify fb_CrtFileCopy()
Instead of doing fseek/ftell/fseek to find out the file size, and then
copy that many bytes, the new version simply copies everything until EOF
is encountered, that's easier and faster.

Besides that, the old code stored the file size in an int and thus was
limited to 32bit, and couldn't handle larger files.