lex: Strictly use dtypes given through number literal suffixes
This changes the lexer to always give a number literal the dtype suggested
by its suffix, if any.

I.e. a number with ll suffix will always be a LONGINT now, and not a
ULONGINT. Or, a number with the % integer suffix will always be an INTEGER
in -lang fb or SHORT in -lang qb now, and never a LONGINT etc.

Most importantly, the l and ul suffixes now properly produce LONGs instead
of being ignored - this is important for 64bit, where numbers with l suffix
should have the 32bit LONG type, while numbers without suffix default to
the 64bit INTEGER type.