parser: Warn about mismatching parameter initializers (closes #438)
This extends hCheckPrototype() to check initializer expressions,
i.e. show a warning if any parameter has different initializers in
prototype and body. Previously the compiler would simply ignore the
2nd initializer and always use the 1st. That can be a problem though,
for example when prototype and body aren't close together, and one
is updated but not the other.

Ideally this would eventually be turned into an error, but that will cause
backwards compatibility issues, and besides that, the new
astIsEqualParamInit() function may need further tweaking before it's 100%
reliable. During development of this patch, at one point, it detected some
false-positives, and even though I fixed those, I'm not sure it is reliable
enough yet, because there are many possible edge cases and tons of test
cases are needed to check them. For an error it'd have to be 100% reliable
though.