parser: Allow quirk function names as literals in PP expressions again
Since the changes to PP expression parsing (starting at 92a83c60),
code such as
	#if typeof(x) = string
failed to work because "string" was treated as a call to the string()
function instead of a literal, causing the compiler to complain about
missing arguments. This caused trouble for quite some people, because
apparently it's a widely used construct...

Even though this isn't good code (it should be <typeof(x) = typeof(string)>
or at least <typeof(x) = "STRING">), it seems wise to re-allow it at least
for quirk functions to improve backwards compatibility. Undeclared
identifiers were also re-allowed in 7ae012df.