C backend: Emit proper source file names also for PRIVATE procedure bodies
Cannot always rely on env.inf.name to point to the source file of the
procedure that's currently being emitted. That's because PRIVATE procedures
are emitted after all others, even behind main(), when env.inf.name points
to the toplevel file.

Luckily we already have symbGetProcIncFile() though to get a procedure's
#include source file name if it came from an #included file. The ASM
backend uses this too.