rtlib: Fix fb_ConPrintTTY() division by zero with tiny graphics screens
If a graphics screen (as created with ScreenRes) is small enough that not
even 1 char (from the built-in gfxlib2 fonts) would fit on it
horizontally, then the width available to fb_ConPrintTTY() as called from
the gfxlib2's PRINT code will be zero. In that case it should do nothing,
and definitely not crash.