symb: Don't specialize some array descriptors to the array's dtype
It was added in 22b784de434ebaf22d75ee8c0bf8dcb114e8d9e8, only for the
C backend, but it doesn't seem to be needed anymore for that.

Anyways, for implementing dynamic array fields, it seems good to avoid
dynamic array descriptors being specialized to the array dtype, because
then we'd need one descriptor for every single dynamic array field. And on
top of that, bydesc parameters already use unspecialized descriptors now,
so extending that to all arrays doesn't seem like a loss. The main
advantage is that we only need to create 1 dynamic array descriptor TYPE
internally, that can be re-used for all dynamic array descriptors.