ast: Don't declare ASM backend's jump tables as array
This prevents an array descriptor from being added for the jump table.

The code in symbAddArrayDesc() which was supposed to prevent the
descriptor from being added didn't work because it checked
symbGetIsJumpTb() on the array, but astBuildJMPTB() only adds this flag
after the symbAddVar() which triggers the symbAddArrayDesc().

(However, it was never emitted, because it was never accessed, so there's
no change in code generation)