ast: Add base offset for descriptor field TYPEINI tree
If the descriptor is a field, then it may have a non-zero offset within
the parent UDT. And since TYPEINIs for fields use absolute offsets (based
on the parent UDT, not relative to individual fields),
astBuildArrayDescIniTree() needs to handle this now too. This is the same
as what's done by cInitializer().