Fix up some cases of Byval As String in examples/headers
For none of these it really mattered though:

For the examples the WinMain() declarations didn't match the Win32 API but
since they're not never used as entry point but instead only used
internally, it didn't really matter.

And the X11 headers were just mistranslated, and are still not ready (won't
even compile), so it doesn't really matter there either.