Unix rtlib: Call pthread_attr_destroy() in fb_ThreadCreate()
Corresponding to every pthread_attr_init() there should be a
pthread_attr_destroy(). Note: pthread_create() doesn't "take ownership"
of the attr object, so the caller is supposed to free it.

I don't know whether this makes any real difference here for us, it also
depends on the pthreads implementation, but it should be safer this way.