parser: Fix #643, typeless REDIM could accidentially re-use field dtype
This fixes the issue where the compiler allowed a typeless REDIM:
	redim x(1 to 2)
where x was not a pre-existing array, but instead a field (visible through
implicit THIS), and then still silently created a new local array using
the field's dtype.

This also adds a ton of test cases for various things related to typeless
REDIMs.