Shrink dynamic array descriptors when dimensions are known at declaration
(see also the previous commit)

If the dimension count of a dynamic array is known at the point of the
declaration, we don't need to allocate a descriptor with room for all of
FB_MAXARRAYDIMS. Instead we can allocate one with the exact amount that
is required, which allows us to save some memory.

Unfortunately it can't be done for arrays declared with (), for example:
	dim array() as integer
which means "unknown dimension count". At the point of the declaration,
the compiler has to assume that the descriptor must have room for
FB_MAXARRAYDIMS, even if we can later determine (thanks to the compile-time
dimension count checking) that it needs less. The problem is that the
descriptor must be allocated and initialized immediately, in most cases,
and can't be changed afterwards anymore. (at least not without rewriting
fbc to be a multi-pass compiler to allow such things)