Overload resolution: Allow passing Long Ptr to Integer Ptr on 32bit, etc.
This adjusts overload resolution to allow more cases of integer pointer
argument being passed to integer pointer parameter. On 32bit, it's ok
to pass a Long Ptr to an Integer Ptr (or vice-versa) because Long and
Integer are basically the same thing. Same goes for LongInt <-> Integer on
64bit.

This is essentially the same as allowing Long arguments to be passed to
Byref As Integer parameters, etc., which already worked.

This helps solving an issue with dir() which also has overloads using such
integer pointers. Since the 64bit port where the parameter of one such
dir() overload was changed from Integer Ptr to Long Ptr, @integer
expressions couldn't be passed anymore. This should still work though.