Automatically link fbgfxmt if gfx and thread functions are used
When using FB graphics and threading functions, we want to use libfbgfxmt
instead of libfbgfx.

Previously in rtl-gfx.bas the compiler just added -lfbgfx and dependencies
directly. But now we need to choose -lfbgfx or -lfbgfxmt depending on
FB_COMPOPT_MULTITHREADED, which may only be set after gfx functions were
seen, for example:
	screen ...
	foo = threadcreate( ... )

Thus we need to delay adding -lfbgfx or -lfbgfxmt because doing it in
rtl-gfx.bas is too early.

Instead, now we have a FB_COMPOPT_GFX (though it's just used internally and
there is no fbc -gfx command line option) which is set in rtl-gfx.bas and
handled by the fbc frontend (hAddDefaultLibs()) and also the objinfo code.