rtlib: Fix ThreadDetach race condition/dangling pointer
Previously, a FBTHREAD context was used by threadcreate() to hold the
information about the user's threadproc and about the created thread.
This FBTHREAD context was passed to the rtlib's internal threadproc so it
could call the user's threadproc, and also returned from threadcreate(),
so threadwait() could wait on it and then free it.

With threaddetach() however, it became possible that the FBTHREAD was freed
in threaddetach() before the rtlib's internal threadproc had a chance to
read it (race condition & dangling pointer).

Thus, this patch splits things up: The information about the created thread
and the information about the user's threadproc is stored into 2 separate
contexts, one of which is threadwait's/threaddetach's responsibility to
free, and the other is our internal threadproc's responsibility to free.

Note: This doesn't change FBTHREAD's sizeof(), because even though it could
now be simplified, it's probably better to do in a separate commit in case
it must be reverted.