rtlib: Fix XBox's threadcreate()
I'm not sure but I think the user data pointer should be passed as-is to
the PsCreateSystemThreadEx() function (like with the thread creation
functions of other platforms), as opposed to passing its address.

Either way it looks like this was never tested since the code in the
threadproc() has been commented out ever since it was added in 64dbdfb1.