ast: Use AST_OPOPT_ISINI with astNewASSIGN() where it makes sense
Everytime astNewASSIGN() is used to "copy-construct" something,
AST_OPOPT_ISINI should be used.

Currently it only matters for UDTs and strings, because for them assignment
and copy construction are different due to constructors and Let overloads,
etc. For simple data types like integers though, AST_OPOPT_ISINI doesn't
make a difference, because assignment & copy construction are the same
thing.

This fixes a case in ast-node-arg.bas, where AST_OPOPT_ISINI was definitely
missing - when passing classes without copy-constructor to a Byval
parameter, the Let overload was being called instead of the implicit
memcpy() (that's currently done in place of the copy-constructor call).