STABS emitter: Overhaul emitting of array descriptor types
This fixes an extra typeAddrOf() that was done when emitting array
descriptor fields, but wasn't needed, and other than that, this just
simplifies how debug info for descriptor types is emitted, hopefully
without introducing regressions.

Most importantly, descriptor types are now emitted just like any other
UDTs, and there's only one HACK remaining to cause the descriptor's dimTB
to be emitted only as big as needed.

This also has the nice side-effect that the FBARRAY.dimTB will now actually
be emitted as an array of type FBARRAYDIM, not as a bunch of weirdly named
fields.