FBARRAY descriptors: Also specialize the ptr field to the array's dtype
It is really the ptr field that holds the buffer containing the array
elements, so it makes sense to give it the array dtype.

In fact, doing this for the ptr field makes more sense than doing it for
the data field, because the data field may not point to the array data
directly (due to the "diff" difference).

http://www.freebasic.net/forum/viewtopic.php?p=197586#p197586