C backend: Emit structs containing bitfields as byte arrays
http://www.freebasic.net/forum/viewtopic.php?f=3&t=21772

df33af45 exposed the problem where structures with bitfields sometimes are
emitted incorrectly enough such that their sizeof() isn't what it should
be.

The core problem here is that FB's bitfields are incompatible to GCC's,
and the 2nd problem is that the C backend didn't do a good job of dealing
with that. This patch simply "improves" it to emit a properly-sized struct
even in such cases - by emitting a byte array in place of the fields.